ai_service/requirements.txt
flask
flask-cors
google-generativeai
spacy
pypdf
python-docx
sentence-transformers
torch
python-dotenv
gtts
gunicorn
